+++ examples/2.1.5.7-grouping-by-query.php
@@ -3,6 +3,8 @@
 require(__DIR__.'/init.php');
 htmlHeader();
 
+echo "<h2>Note: Query grouping seems to be broken in Solr 8.0.0!</h2>";
+
 // create a client instance
 $client = new Solarium\Client($config);