+++ examples/2.3.1-mlt-query.php
@@ -9,38 +9,30 @@ htmlHeader();
 $client = new Client($config);
 
 // get a morelikethis query instance
-$query = $client->createMoreLikeThis();
+$query = $client->createSelect()
+    // Unfortunately the /mlt handler of the techproducts examlpe doesn't exist anymore.
+    // Therefore we have to use the /browse handler and turn of velocity by forcing json as response writer.
+    ->setHandler('browse')
+    ->setResponseWriter(\Solarium\Core\Query\AbstractQuery::WT_JSON);
 
-$query->setQuery('id:SP2514N');
-$query->setMltFields('manu,cat');
-$query->setMinimumDocumentFrequency(1);
-$query->setMinimumTermFrequency(1);
-$query->createFilterQuery('stock')->setQuery('inStock:true');
-$query->setInterestingTerms('details');
-$query->setMatchInclude(true);
+$query->getMoreLikeThis()
+    ->setFields('manu,cat')
+    ->setMinimumDocumentFrequency(1)
+    ->setMinimumTermFrequency(1)
+    ->setInterestingTerms('details')
+    ->setMatchInclude(true);
+
+$query->setQuery('id:SP2514N')
+    ->createFilterQuery('stock')->setQuery('inStock:true');
 
 // this executes the query and returns the result
 $resultset = $client->select($query);
 
-echo 'Document used for matching:<br/><table>';
-foreach ($resultset->getMatch() as $field => $value) {
-    // this converts multivalue fields to a comma-separated string
-    if (is_array($value)) {
-        $value = implode(', ', $value);
-    }
-
-    echo '<tr><th>' . $field . '</th><td>' . $value . '</td></tr>';
-}
-echo '</table><hr/>';
-
-// display the total number of MLT documents found by solr
-echo 'Number of MLT matches found: '.$resultset->getNumFound().'<br/><br/>';
-echo '<b>Listing of matched docs:</b>';
-
-// show MLT documents using the resultset iterator
+echo 'Documents used for matching:<br/>';
+// show documents using the resultset iterator
 foreach ($resultset as $document) {
 
-    echo '<hr/><table>';
+    echo '<table>';
 
     // the documents are also iterable, to get all fields
     foreach ($document as $field => $value) {
@@ -53,6 +45,33 @@ foreach ($resultset as $document) {
     }
 
     echo '</table>';
+}
+
+echo '<hr/>';
+
+$mlt = $resultset->getMoreLikeThis();
+
+// display the total number of MLT documents found by solr
+echo 'Number of MLT matches found: '.$resultset->getNumFound().'<br/><br/>';
+echo '<b>Listing of matched docs:</b>';
+
+// show MLT documents using the resultset iterator
+foreach ($mlt as $results) {
+    foreach ($results as $document) {
+        echo '<hr/><table>';
+
+        // the documents are also iterable, to get all fields
+        foreach ($document as $field => $value) {
+            // this converts multivalue fields to a comma-separated string
+            if (is_array($value)) {
+                $value = implode(', ', $value);
+            }
+
+            echo '<tr><th>' . $field . '</th><td>' . $value . '</td></tr>';
+        }
+
+        echo '</table>';
+    }
 }
 
 htmlFooter();