+++ examples/init.php
@@ -5,19 +5,16 @@ ini_set('display_errors', true);
 
 if (file_exists($_SERVER['DOCUMENT_ROOT'] .'/config.php')) {
     require($_SERVER['DOCUMENT_ROOT'] .'/config.php');
-    require $config['autoload'];
-} else {
-    require('config.dist.php');
-    require __DIR__.'/vendor/autoload.php';
 }
 
+require $config['autoload'] ?? __DIR__.'/../vendor/autoload.php';
 
 function htmlHeader()
 {
-    echo '<html><head><title>Solarium examples</title></head><body>';
+    echo '<html><head><title>Solarium examples</title></head><body><nav><a href="index.html">Back to Overview</a></nav><br><article>';
 }
 
 function htmlFooter()
 {
-    echo '</body></html>';
+    echo '</article><br><nav><a href="index.html">Back to Overview</a></nav></body></html>';
 }